refactor!: drop support for legacy drawer implementation
