+++ packages/react-native-drawer-layout/src/constants.tsx
@@ -1,4 +1,4 @@
+export const SWIPE_EDGE_WIDTH = 32;
 export const SWIPE_MIN_OFFSET = 5;
 export const SWIPE_MIN_DISTANCE = 60;
 export const SWIPE_MIN_VELOCITY = 500;
-export const DEFAULT_DRAWER_WIDTH = '80%';
